+++ packages/devui-vue/devui/form/src/components/form-control/form-control.tsx
@@ -1,7 +1,7 @@
 import { defineComponent, inject, ref, computed, reactive, onMounted, Teleport } from 'vue';
 import { uniqueId } from 'lodash';
 import { IForm, formControlProps, formInjectionKey } from '../../form-types';
-import { ShowPopoverErrorMessageEventData } from '../directive/d-validate-rules';
+import { ShowPopoverErrorMessageEventData } from '../../directives/d-validate-rules';
 import clickoutsideDirective from '../../../../shared/devui-directive/clickoutside';
 import { EventBus, getElOffset } from '../../utils';
 import Icon from '../../../../icon/src/icon';
