+++ packages/devui-vue/devui/form/src/directives/d-validate-rules.ts
@@ -1,7 +1,7 @@
 import AsyncValidator, { RuleItem } from 'async-validator';
 import { VNode, DirectiveBinding } from 'vue';
 import { debounce } from 'lodash';
-import { EventBus, isObject, hasKey } from '../../utils';
+import { EventBus, isObject, hasKey } from '../utils';
 import './style.scss';
 
 interface ValidateFnParam {
